requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.10
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
docplex==2.29.241
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.3.0
grpcio==1.70.0
gTTS==2.5.4
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipympl==0.9.7
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
Js2Py==0.74
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.1.1
ml_dtypes==0.5.1
more-itertools==10.6.0
MouseInfo==0.1.3
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
noisereduce==3.0.3
notebook==7.3.2
notebook_shim==0.2.4
Nuitka==2.6.9
numba==0.61.0
numpy==1.26.4
nvidia-pyindex==1.0.9
oauthlib==3.2.2
onnxruntime-gpu==1.20.1
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==11.1.0
pipwin==0.5.2
platformdirs==4.3.6
playsound==1.3.0
ply==3.11
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.20.3
psutil==7.0.0
PuLP==3.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybind11==2.13.6
pycaw==20240210
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyomo==6.9.2
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.2.1
python-telegram-bot==22.0
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.28.1
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
tflite==2.18.0
tflite-support==0.1.0a1
thinc==8.3.4
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torch-cluster==1.6.3+pt20cu118
torch-scatter==2.1.2+pt20cu118
torch-sparse==0.6.18+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.0
translate==3.6.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.98
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
vosk==0.3.45
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
whisper==1.1.10
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xx_ent_wiki_sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl#sha256=6f3c4b853852ea9e9d2dc76cc950dddb10a7e4c42d813308caefe6c5e8be2f0a
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
